pyproject.toml
[project]
name = "fastapi-todo-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bcrypt==4.0.1",
    "fastapi[standard]>=0.135.1",
    "passlib>=1.7.4",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.48",
]
